// src/shared/config/swagger/types/models/CreateBrandDto.ts
/**
 * Internet store application
 * The Internet store API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateBrandDto {
    'name': string;
    'websiteUrl': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "websiteUrl",
            "baseName": "websiteUrl",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateBrandDto.attributeTypeMap;
    }

    public constructor() {
    }
}