// src/shared/config/swagger/types/models/CreateCommentReactionDto.ts
/**
 * Internet store application
 * The Internet store API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateCommentReactionDto {
    'reaction': CreateCommentReactionDtoReactionEnum;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "reaction",
            "baseName": "reaction",
            "type": "CreateCommentReactionDtoReactionEnum",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateCommentReactionDto.attributeTypeMap;
    }

    public constructor() {
    }
}


export type CreateCommentReactionDtoReactionEnum = "LIKE" | "DISLIKE" ;