// src/shared/config/swagger/types/models/CreateItemDto.ts
/**
 * Internet store application
 * The Internet store API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateItemDto {
    'title': string;
    'price': number;
    'discountPercentage': number;
    'typeId': string;
    'brandId': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "title",
            "baseName": "title",
            "type": "string",
            "format": ""
        },
        {
            "name": "price",
            "baseName": "price",
            "type": "number",
            "format": ""
        },
        {
            "name": "discountPercentage",
            "baseName": "discountPercentage",
            "type": "number",
            "format": ""
        },
        {
            "name": "typeId",
            "baseName": "type_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "brandId",
            "baseName": "brand_id",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateItemDto.attributeTypeMap;
    }

    public constructor() {
    }
}