// src/shared/config/swagger/types/models/EditUserDto.ts
/**
 * Internet store application
 * The Internet store API description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class EditUserDto {
    'userName': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "userName",
            "baseName": "userName",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return EditUserDto.attributeTypeMap;
    }

    public constructor() {
    }
}